Initial refactor and inital unit tests but training code has not been tested
