--- experiments/20250418_150225_Baseline_60m/README.md
@@ -1,2 +1,0 @@
-# 60m parameter baseline
-trained on ~.1b tokens (seed = 42) on a 3070 with 8gb of vram that cost $0.113/hr
