--- experiments/20250418_190215_baseline_120m/requirements.txt
@@ -1,7 +1,0 @@
-datasets==3.5.0
-numpy==2.2.4
-regex==2024.11.6
-Requests==2.32.3
-tiktoken==0.9.0
-torch==2.6.0
-tqdm==4.66.5
