--- experiments/20250418_204839_Baseline_343m/download_hellaswag.txt
@@ -1,66 +1,0 @@
-"""
-Downloads and evaluates HellaSwag in Python.
-https://github.com/rowanz/hellaswag
-
-Example HellaSwag json item:
-
-{"ind": 24, "activity_label": "Roof shingle removal", "ctx_a": "A man is sitting on a roof.", "ctx_b": "he", "ctx": "A man is sitting on a roof. he", "split": "val", "split_type": "indomain", "label": 3, "endings": ["is using wrap to wrap a pair of skis.", "is ripping level tiles off.", "is holding a rubik's cube.", "starts pulling up roofing on a roof."], "source_id": "activitynet~v_-JhWjGDPHMY"}
-
-ind: dataset ID
-activity_label: The ActivityNet or WikiHow label for this example
-context: There are two formats. The full context is in ctx. When the context ends in an (incomplete) noun phrase, like for ActivityNet, this incomplete noun phrase is in ctx_b, and the context up until then is in ctx_a. This can be useful for models such as BERT that need the last sentence to be complete. However, it's never required. If ctx_b is nonempty, then ctx is the same thing as ctx_a, followed by a space, then ctx_b.
-endings: a list of 4 endings. The correct index is given by label (0,1,2, or 3)
-split: train, val, or test.
-split_type: indomain if the activity label is seen during training, else zeroshot
-source_id: Which video or WikiHow article this example came from
-
-gpt2 (124M)
-- eleuther harness reports acc 28.92%, acc_norm 31.14% (multiple choice style)
-- this script: 10042 acc: 0.2859 acc_norm: 0.2955 (completion style)
-
-gpt2-xl (1558M)
-- eleuther harness reports acc 40.04%, acc_norm 50.89% (multiple choice style)
-- this script: 10042 acc: 0.3842 acc_norm: 0.4893 (completion style)
-
-The validation set of HellaSwag has a total of 10,042 examples.
-"""
-
-import os
-import json
-import requests
-from tqdm import tqdm
-
-# -----------------------------------------------------------------------------
-DATA_CACHE_DIR = os.path.join(os.path.dirname(__file__), 'data')
-
-def download_file(url: str, fname: str, chunk_size=1024):
-    """Helper function to download a file from a given url"""
-    resp = requests.get(url, stream=True)
-    total = int(resp.headers.get("content-length", 0))
-    with open(fname, "wb") as file, tqdm(
-        desc=fname,
-        total=total,
-        unit="iB",
-        unit_scale=True,
-        unit_divisor=1024,
-    ) as bar:
-        for data in resp.iter_content(chunk_size=chunk_size):
-            size = file.write(data)
-            bar.update(size)
-
-hellaswags = {
-    "train": "https://raw.githubusercontent.com/rowanz/hellaswag/master/data/hellaswag_train.jsonl",
-    "val": "https://raw.githubusercontent.com/rowanz/hellaswag/master/data/hellaswag_val.jsonl",
-    "test": "https://raw.githubusercontent.com/rowanz/hellaswag/master/data/hellaswag_test.jsonl",
-}
-
-def download(split):
-    """Downloads HellaSwag DATA_CACHE_DIR"""
-    os.makedirs(DATA_CACHE_DIR, exist_ok=True)
-    data_url = hellaswags[split]
-    data_filename = os.path.join(DATA_CACHE_DIR, f"hellaswag_{split}.jsonl")
-    if not os.path.exists(data_filename):
-        print(f"Downloading {data_url} to {data_filename}...")
-        download_file(data_url, data_filename)
-
-download('val')
