+++ pyproject.toml
@@ -1,0 +1,30 @@
+[project]
+name = "core-gpt"
+version = "0.1.0"
+description = "Add your description here"
+readme = "README.md"
+requires-python = ">=3.10"
+dependencies = [
+    "datasets==3.5.0",
+    "hydra-core>=1.3.2",
+    "numpy==2.2.4",
+    "omegaconf>=2.3.0",
+    "pytest>=8.3.5",
+    "pytest-cov>=6.1.1",
+    "regex==2024.11.6",
+    "requests==2.32.3",
+    "tiktoken==0.9.0",
+    "torch==2.6.0",
+    "tqdm==4.66.5",
+]
+
+[tool.pytest.ini_options]
+addopts = [
+    "--import-mode=importlib",
+]
+pythonpath = [
+  "."
+]
+testpaths = ["tests"]
+python_files = "test_*.py"
+python_functions = "test_*"
